requirements.txt
django>=4.2

Pillow>=10.0
python-decouple>=3.8
certifi>=2024.0.0
